filezup_backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
